+++ synphot/compat.py
@@ -2,6 +2,7 @@
 """Module to handle backward-compatibility."""
 
 import astropy
+import numpy
 from astropy.utils.introspection import minversion
 
 try:
@@ -18,11 +19,12 @@ except ImportError:
 else:
     HAS_DUST_EXTINCTION = True
 
-
 __all__ = ['ASTROPY_LT_5_0', 'ASTROPY_LT_4_3', 'ASTROPY_LT_4_1',
-           'ASTROPY_LT_4_0', 'HAS_SPECUTILS', 'HAS_DUST_EXTINCTION']
+           'ASTROPY_LT_4_0', 'NUMPY_LT_1_17', 'HAS_SPECUTILS',
+           'HAS_DUST_EXTINCTION']
 
 ASTROPY_LT_5_0 = not minversion(astropy, '4.99')  # astropy<5 but includes 5.0.dev  # noqa
 ASTROPY_LT_4_3 = not minversion(astropy, '4.3')
 ASTROPY_LT_4_1 = not minversion(astropy, '4.1')
 ASTROPY_LT_4_0 = not minversion(astropy, '4.0')
+NUMPY_LT_1_17 = not minversion(numpy, '1.17')