+++ synphot/filter_parameterization/__init__.py
@@ -1,0 +1,9 @@
+"""This subpackage handles filter parameterization.
+
+The algorithms in this subpackage were originally developed by
+Brett Morris as part of the `tynt <https://github.com/bmorris3/tynt>`_
+package.
+
+"""
+
+from .filter_fft import *  # noqa